requirements.txt
Django==2.2.3
djangorestframework==3.9.4
factory-boy==2.12.0
pytest==5.0.1
pytest-django==3.5.1
python-dateutil==2.8.0